+++ Domain/Interfaces/IUserRepository.cs
@@ -9,7 +9,7 @@ namespace Domain.Interfaces
 {
     public interface IUserRepository
     {
-        public Task<User?> GetUserById(int id);
+        public Task<User?> GetUserById(long id);
         public Task InsertUser(User user);
         public Task<User?> GetUserByEmail(string email);
         public Task<User?> GetUserByPhoneNumber(string phone);
